requirements.txt
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==2.2
django-allauth==0.42.0
django-cors-headers==3.5.0
django-filter==2.3.0
django-rest-auth==0.9.5
djangorestframework==3.11.1
idna==2.10
importlib-metadata==1.7.0
Markdown==3.2.2
oauthlib==3.1.0
Pillow==7.2.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
zipp==3.1.0

django-nested-admin
coreapi

django-heroku
whitenoise
gunicorn